refactor: move status utility to kubernetes backend and rename
