+++ docs/source/train/custom-training.rst
@@ -126,25 +126,25 @@ variables into your training pods and you can report real-time progress:
 .. code-block:: python
 
    import time
-   from kubeflow.trainer.utils import update_runtime_status
+   from kubeflow.trainer import update_trainjob_status
 
    def train():
        total_steps = 1000
        start_time = time.time()
-       update_runtime_status(progress_percent=0, force=True)  # Training start
+       update_trainjob_status(progress_percent=0, force=True)  # Training start
 
        for step in range(total_steps):
            loss = train_step(batch)
            progress = int((step / total_steps) * 100)
            elapsed = time.time() - start_time
            eta = int((elapsed / (step + 1)) * (total_steps - step - 1)) if step > 0 else None
-           update_runtime_status(
+           update_trainjob_status(
                progress_percent=progress,
                estimated_time_remaining=eta,
                metrics={"loss": loss, "step": step},
            )  # SDK throttles to max 1 update per 5 seconds
 
-       update_runtime_status(progress_percent=100, force=True)  # Training end
+       update_trainjob_status(progress_percent=100, force=True)  # Training end
 
 This is safe to call in any environment — it returns ``False`` silently if not running
 inside a Kubeflow TrainJob with the feature gate enabled.
